Reorganize file structure
1. Put all protobuf definitions in one file
2. rename all BUILD and WORKSPACE file to .bazel
3. put the p4-sample programs in a separate directory under / since they will be shared by many components
4. put bmv2 json parsing cc file, protobufs, and test files, including expected files and python script, under src/bmv2
5. Make appropriate changes to package naming, cc includes and bazel loads, and test scripts
6. Re-write and improve veth_setup.sh and update p4-samples Makefiles.
