+++ p4-samples/hardcoded/.gitignore
