+++ p4-samples/hardcoded/Makefile
@@ -14,17 +14,7 @@
 
 SHELL := /bin/bash
 
-run: build simple_switch control_plane packets
-
-#
-# Control Plane
-# insert entries into p4 tables
-#
-define CONTROL_PLANE_COMMANDS =
-table_add ports_exact set_egress_spec 0 => 1\n\
-table_add ports_exact set_egress_spec 1 => 0\n\
-table_dump ipv4_lpm
-endef
+run: build simple_switch packets
 
 #
 # Sending Test Packets
@@ -42,6 +32,8 @@ sendp(p, iface="veth1")\n\
 q = Ether()/IP(dst="20.20.0.8")/UDP()\n\
 sendp(q, iface="veth3")\n
 endef
+
+VETH_PAIRS_COUNT := 2
 
 # include main Makefile with rules
 include ../Makefile
