+++ p4-samples/reflector/Makefile
@@ -14,33 +14,27 @@
 
 SHELL := /bin/bash
 
-run: build simple_switch control_plane packets
-
-#
-# Control Plane
-# insert entries into p4 tables
-#
-define CONTROL_PLANE_COMMANDS =
-table_add ipv4_lpm ipv4_forward 10.10.0.0/16 => 00:00:00:00:00:00 0\n\
-table_add ipv4_lpm ipv4_forward 20.20.0.0/16 => 22:00:00:00:00:22 1\n\
-table_dump ipv4_lpm
-endef
+run: build simple_switch packets
 
 #
 # Sending Test Packets
 #
 define PACKETS =
+# These wont show up \n\
 p = Ether()/IP(dst="20.20.0.1")/UDP()\n\
 sendp(p, iface="veth1")\n\
 p = Ether()/IP(dst="20.20.0.2")/UDP()\n\
 sendp(p, iface="veth1")\n\
-# this should show up once \n\
-p = Ether()/IP(dst="10.10.0.1")/UDP()\n\
+# These will show up twice each \n\
+p = Ether()/IP(dst="20.20.0.3")/UDP()\n\
 sendp(p, iface="veth3")\n\
-# this should not show up \n\
-p = Ether()/IP(dst="10.10.0.2")/UDP()\n\
-sendp(p, iface="veth1")\n
+p = Ether()/IP(dst="20.20.0.4")/UDP()\n\
+sendp(p, iface="veth3")\n\
+q = Ether()/IP(dst="20.20.0.8")/UDP()\n\
+sendp(q, iface="veth3")\n
 endef
+
+VETH_PAIRS_COUNT := 2
 
 # include main Makefile with rules
 include ../Makefile
