--- src/BUILD
@@ -1,27 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# This file contains build rules for the main
-# executable binary in the project.
-
-load("@rules_cc//cc:defs.bzl", "cc_binary")
-
-cc_binary(
-    name = "main",
-    srcs = ["main.cc"],
-    deps = [
-        "//src/protobuf/bmv2:p4_parsing_cc_proto"
-    ],
-    visibility = ["//test:__pkg__"]
-)
