+++ src/bmv2/expected/reflector.pb.txt
@@ -214,7 +214,7 @@ parsers {
 deparsers {
   name: "deparser"
   source_info {
-    filename: "test/samples/reflector/reflector.p4"
+    filename: "p4-samples/reflector/reflector.p4"
     line: 45
     column: 8
     source_fragment: "UselessDeparser"
@@ -295,7 +295,7 @@ actions {
           fields {
             key: "filename"
             value {
-              string_value: "test/samples/reflector/reflector.p4"
+              string_value: "p4-samples/reflector/reflector.p4"
             }
           }
           fields {
@@ -318,7 +318,7 @@ actions {
 pipelines {
   name: "ingress"
   source_info {
-    filename: "test/samples/reflector/reflector.p4"
+    filename: "p4-samples/reflector/reflector.p4"
     line: 50
     column: 8
     source_fragment: "MyIngress"
@@ -327,7 +327,7 @@ pipelines {
   tables {
     name: "tbl_act"
     source_info {
-      filename: "test/samples/reflector/reflector.p4"
+      filename: "p4-samples/reflector/reflector.p4"
       line: 54
       column: 38
       source_fragment: "="
@@ -343,7 +343,7 @@ pipelines {
   name: "egress"
   id: 1
   source_info {
-    filename: "test/samples/reflector/reflector.p4"
+    filename: "p4-samples/reflector/reflector.p4"
     line: 35
     column: 8
     source_fragment: "UselessEgress"
