+++ src/bmv2/expected/table.pb.txt
@@ -214,103 +214,22 @@ parsers {
 deparsers {
   name: "deparser"
   source_info {
-    filename: "test/samples/hardcoded/hardcoded.p4"
+    filename: "p4-samples/port-table/table.p4"
     line: 45
     column: 8
     source_fragment: "UselessDeparser"
   }
 }
 actions {
-  name: "act"
-  primitives {
-    fields {
-      key: "op"
-      value {
-        string_value: "assign"
-      }
-    }
-    fields {
-      key: "parameters"
-      value {
-        list_value {
-          values {
-            struct_value {
-              fields {
-                key: "type"
-                value {
-                  string_value: "field"
-                }
-              }
-              fields {
-                key: "value"
-                value {
-                  list_value {
-                    values {
-                      string_value: "standard_metadata"
-                    }
-                    values {
-                      string_value: "egress_spec"
-                    }
-                  }
-                }
-              }
-            }
-          }
-          values {
-            struct_value {
-              fields {
-                key: "type"
-                value {
-                  string_value: "hexstr"
-                }
-              }
-              fields {
-                key: "value"
-                value {
-                  string_value: "0x0001"
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    fields {
-      key: "source_info"
-      value {
-        struct_value {
-          fields {
-            key: "column"
-            value {
-              number_value: 12
-            }
-          }
-          fields {
-            key: "filename"
-            value {
-              string_value: "test/samples/hardcoded/hardcoded.p4"
-            }
-          }
-          fields {
-            key: "line"
-            value {
-              number_value: 55
-            }
-          }
-          fields {
-            key: "source_fragment"
-            value {
-              string_value: "standard_metadata.egress_spec = 1"
-            }
-          }
-        }
-      }
-    }
-  }
+  name: "NoAction"
 }
 actions {
-  name: "act_0"
+  name: "MyIngress.set_egress_spec"
   id: 1
+  runtime_data {
+    name: "port"
+    bitwidth: 9
+  }
   primitives {
     fields {
       key: "op"
@@ -350,13 +269,13 @@ actions {
               fields {
                 key: "type"
                 value {
-                  string_value: "hexstr"
+                  string_value: "runtime_data"
                 }
               }
               fields {
                 key: "value"
                 value {
-                  string_value: "0x0000"
+                  number_value: 0
                 }
               }
             }
@@ -371,25 +290,25 @@ actions {
           fields {
             key: "column"
             value {
-              number_value: 12
+              number_value: 8
             }
           }
           fields {
             key: "filename"
             value {
-              string_value: "test/samples/hardcoded/hardcoded.p4"
+              string_value: "p4-samples/port-table/table.p4"
             }
           }
           fields {
             key: "line"
             value {
-              number_value: 57
+              number_value: 56
             }
           }
           fields {
             key: "source_fragment"
             value {
-              string_value: "standard_metadata.egress_spec = 0"
+              string_value: "standard_metadata.egress_spec = port"
             }
           }
         }
@@ -400,47 +319,40 @@ actions {
 pipelines {
   name: "ingress"
   source_info {
-    filename: "test/samples/hardcoded/hardcoded.p4"
+    filename: "p4-samples/port-table/table.p4"
     line: 50
     column: 8
     source_fragment: "MyIngress"
   }
-  init_table: "node_2"
+  init_table: "MyIngress.ports_exact"
   tables {
-    name: "tbl_act"
+    name: "MyIngress.ports_exact"
     source_info {
-      filename: "test/samples/hardcoded/hardcoded.p4"
-      line: 55
-      column: 42
-      source_fragment: "="
+      filename: "p4-samples/port-table/table.p4"
+      line: 60
+      column: 10
+      source_fragment: "ports_exact"
     }
-    match_type: "exact"
-    type: "simple"
-    max_size: 1024
-    action_ids: 0
-    actions: "act"
-  }
-  tables {
-    name: "tbl_act_0"
-    id: 1
-    source_info {
-      filename: "test/samples/hardcoded/hardcoded.p4"
-      line: 57
-      column: 42
-      source_fragment: "="
+    key {
+      match_type: "exact"
+      name: "standard_metadata.ingress_port"
+      target: "standard_metadata"
+      target: "ingress_port"
     }
     match_type: "exact"
     type: "simple"
     max_size: 1024
     action_ids: 1
-    actions: "act_0"
+    action_ids: 0
+    actions: "MyIngress.set_egress_spec"
+    actions: "NoAction"
   }
 }
 pipelines {
   name: "egress"
   id: 1
   source_info {
-    filename: "test/samples/hardcoded/hardcoded.p4"
+    filename: "p4-samples/port-table/table.p4"
     line: 35
     column: 8
     source_fragment: "UselessEgress"
