+++ src/bmv2/test_sdiff.py
