--- src/protobuf/bmv2/BUILD
@@ -1,39 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Contains the build rules to generate cc files
-# from the protobuf definitions in this directory
-
-load("@rules_cc//cc:defs.bzl", "cc_proto_library")
-load("@rules_proto//proto:defs.bzl", "proto_library")
-
-cc_proto_library(
-    name = "p4_parsing_cc_proto",
-    deps = [":p4_parsing_proto"],
-    visibility = ["//src:__pkg__"]
-)
-
-proto_library(
-    name = "p4_parsing_proto",
-    srcs = [
-      "header.proto",
-      "parser.proto",
-      "primitives.proto",
-      "program.proto",
-      "table.proto"
-    ],
-    deps = [
-      "@com_google_protobuf//:struct_proto",
-    ]
-)
